Added closing color string
